Move *All functions to chart-registry module
